Crate compose_error_codes_doc_tests

Documentation and doc-tests for the Compose error codes.

The purpose of this crate is to provide the compose error codes as a Rust module, so that they can be included in docs.rs and the examples can be tested.

The source of the error codes is in the compose-error-codes crate. The description of the error code is transformed via compose_doc::transform_markdown into documented items with doc-tests.

Modules

DOC_E0001_UNCLOSED_DELIMITER

E0001: Unclosed delimiter

DOC_E0002_INVALID_ASSIGNMENT

E0002: Assignments are not allowed in expression contexts

DOC_E0003_EXPECTED_BINDING_AFTER_LET

E0003: Expected binding after let

DOC_E0004_MUTATE_IMMUTABLE_VARIABLE

E0004: Cannot reassign to a variable declared as immutable

DOC_E0005_IF_EXPRESSION_BODIES_REQUIRE_BRACES

E0005: if expressions must use braces to delimit their body.

DOC_E0006_UNTERMINATED_STATEMENT

E0006: A semicolon is required after each statement.

DOC_E0007_MISSING_EQUALS_AFTER_LET_BINDING

E0007: A let binding must be followed by = or terminated with a semicolon or newline.

DOC_E0008_EXPECTED_EXPRESSION

E0008: An expression was expected but none was found

DOC_E0009_ARGS_MISSING_COMMAS

E0009: Function arguments must be separated by commas

DOC_E0010_UNCAPTURED_VARIABLE

E0010: Outer variables used in closure but not captured

DOC_E0011_UNBOUND_VARIABLE

E0010: Unbound variable

DOC_E0012_PREDICATE_MUST_RETURN_BOOLEAN

E0012: Predicate functions must return a boolean value

DOC_E0301_ARRAY_DESTRUCTURING_WRONG_NUMBER_OF_ELEMENTS

E0301: Array destructuring pattern did not fit array length

DOC_E0302_MAP_DESTRUCTURING_UNCOVERED_KEYS

E0302: map destructuring does not cover all keys

DOC_E0303_MAP_DESTRUCTURING_MISSING_KEY_IN_VALUE

E0303: missing key in map pattern

DOC_E0304_TYPE_PATTERNS_NOT_ALLOWED_IN_LET_BINDINGS

E0304: Type patterns are not allowed in let bindings

DOC_E0311_MATCH_ARM_PATTERNS_BIND_DIFFERENT_VARIABLES

E0311: Match arm patterns bind different variables

DOC_W0001_USED_UNINITIALIZED_VARIABLE

W0001: Use of a variable before it has been initialized